Add initial functionality for Plusterm wrapper on xterm.js
This commit introduces the initial setup for the Plusterm project, including a new package.json and package-lock.json with xterm.js as a dependency. A basic implementation for a wrapper on xterm.js has been created in src/index.js file. Besides, a demo has been provided in demo/index.html to demonstrate the functionalities. The main aim of this commit is to establish the structure of the project and lay out the basic functionality. More features and improvements will be added progressively.